+++ docs/components-reference.md
@@ -120,9 +120,10 @@ import { Pagination } from '@/components/controls'
 
 ## typography
 
-| ディレクトリ  | 説明                                                                                          |
-| ------------- | --------------------------------------------------------------------------------------------- |
-| `typography/` | テキスト表示コンポーネント。`Heading`（h1/h2/h3 バリアント対応）、`Paragraph`（段落）を提供。 |
+| ディレクトリ            | 説明                                                    |
+| ----------------------- | ------------------------------------------------------- |
+| `typography/headings`   | 見出しコンポーネント。`Heading` を提供。h1/h2/h3 対応。 |
+| `typography/paragraphs` | 段落コンポーネント。`Paragraph` を提供。                |
 
 ## ui
