+++ src/components/typography/paragraphs/index.ts
@@ -1,2 +1,1 @@
 export { default as Paragraph } from './paragraph'
-export { default as Heading } from './heading'
